Moved Swarm Peer from node API to plum pork
