+++ api/api.go
@@ -4,5 +4,4 @@ package api
 // API is the user interface to a Filecoin node.
 type API interface {
 	Daemon() Daemon
-	Swarm() Swarm
 }
