+++ api/impl/errors.go
@@ -9,6 +9,4 @@ var (
 	// "from" address and we couldn't default it because there were zero or
 	// more than one to choose from.
 	ErrCouldNotDefaultFromAddress = errors.New("no from address specified and no default address available")
-	// ErrNodeOffline indicates that the node must not be offline for the operation performed.
-	ErrNodeOffline = errors.New("node must be online")
 )
