--- api/impl/swarm.go
@@ -1,68 +1,0 @@
-package impl
-
-import (
-	"context"
-	"sort"
-
-	"github.com/filecoin-project/go-filecoin/api"
-)
-
-type nodeSwarm struct {
-	api *nodeAPI
-}
-
-// COPIED FROM go-ipfs core/commands/swarm.go
-// TODO a lot of this functionality should migrate to the network package.
-
-func newNodeSwarm(api *nodeAPI) *nodeSwarm {
-	return &nodeSwarm{api: api}
-}
-
-func (ns *nodeSwarm) Peers(ctx context.Context, verbose, latency, streams bool) (*api.SwarmConnInfos, error) {
-	nd := ns.api.node
-
-	if nd.Host() == nil {
-		return nil, ErrNodeOffline
-	}
-
-	conns := nd.Host().Network().Conns()
-
-	var out api.SwarmConnInfos
-	for _, c := range conns {
-		pid := c.RemotePeer()
-		addr := c.RemoteMultiaddr()
-
-		ci := api.SwarmConnInfo{
-			Addr: addr.String(),
-			Peer: pid.Pretty(),
-		}
-
-		/* FIXME(steb)
-		swcon, ok := c.(*swarm.Conn)
-		if ok {
-			ci.Muxer = fmt.Sprintf("%T", swcon.StreamConn().Conn())
-		}
-		*/
-
-		if verbose || latency {
-			lat := nd.Host().Peerstore().LatencyEWMA(pid)
-			if lat == 0 {
-				ci.Latency = "n/a"
-			} else {
-				ci.Latency = lat.String()
-			}
-		}
-		if verbose || streams {
-			strs := c.GetStreams()
-
-			for _, s := range strs {
-				ci.Streams = append(ci.Streams, api.SwarmStreamInfo{Protocol: string(s.Protocol())})
-			}
-		}
-		sort.Sort(&ci)
-		out.Peers = append(out.Peers, ci)
-	}
-
-	sort.Sort(&out)
-	return &out, nil
-}
