+++ net/network.go
@@ -3,6 +3,7 @@ package net
 import (
 	"context"
 	"fmt"
+	"sort"
 
 	ma "gx/ipfs/QmNTCey11oxhb1AxDnQBRHtdhap6Ctud872NjAYPYYXPuc/go-multiaddr"
 	"gx/ipfs/QmTu65MVbemtUxJEWgsTtzv9Zv9P8rvmqNA4eG9TrTRGYc/go-libp2p-peer"
@@ -20,6 +21,49 @@ import (
 type SwarmConnectResult struct {
 	Peer    string
 	Success bool
+}
+
+// SwarmConnInfo represents details about a single swarm connection.
+type SwarmConnInfo struct {
+	Addr    string
+	Peer    string
+	Latency string
+	Muxer   string
+	Streams []SwarmStreamInfo
+}
+
+// SwarmStreamInfo represents details about a single swarm stream.
+type SwarmStreamInfo struct {
+	Protocol string
+}
+
+func (ci *SwarmConnInfo) Less(i, j int) bool {
+	return ci.Streams[i].Protocol < ci.Streams[j].Protocol
+}
+
+func (ci *SwarmConnInfo) Len() int {
+	return len(ci.Streams)
+}
+
+func (ci *SwarmConnInfo) Swap(i, j int) {
+	ci.Streams[i], ci.Streams[j] = ci.Streams[j], ci.Streams[i]
+}
+
+// SwarmConnInfos represent details about a list of swarm connections.
+type SwarmConnInfos struct {
+	Peers []SwarmConnInfo
+}
+
+func (ci SwarmConnInfos) Less(i, j int) bool {
+	return ci.Peers[i].Addr < ci.Peers[j].Addr
+}
+
+func (ci SwarmConnInfos) Len() int {
+	return len(ci.Peers)
+}
+
+func (ci SwarmConnInfos) Swap(i, j int) {
+	ci.Peers[i], ci.Peers[j] = ci.Peers[j], ci.Peers[i]
 }
 
 // Network is a unified interface for dealing with libp2p
@@ -90,4 +134,52 @@ func (network *Network) Connect(ctx context.Context, addrs []string) ([]SwarmCon
 	}
 
 	return output, nil
+}
+
+// Peers lists peers currently available on the network
+func (network *Network) Peers(ctx context.Context, verbose, latency, streams bool) (*SwarmConnInfos, error) {
+	if network.host == nil {
+		return nil, errors.New("node must be online")
+	}
+
+	conns := network.host.Network().Conns()
+
+	var out SwarmConnInfos
+	for _, c := range conns {
+		pid := c.RemotePeer()
+		addr := c.RemoteMultiaddr()
+
+		ci := SwarmConnInfo{
+			Addr: addr.String(),
+			Peer: pid.Pretty(),
+		}
+
+		/* FIXME(steb)
+		swcon, ok := c.(*swarm.Conn)
+		if ok {
+			ci.Muxer = fmt.Sprintf("%T", swcon.StreamConn().Conn())
+		}
+		*/
+
+		if verbose || latency {
+			lat := network.host.Peerstore().LatencyEWMA(pid)
+			if lat == 0 {
+				ci.Latency = "n/a"
+			} else {
+				ci.Latency = lat.String()
+			}
+		}
+		if verbose || streams {
+			strs := c.GetStreams()
+
+			for _, s := range strs {
+				ci.Streams = append(ci.Streams, SwarmStreamInfo{Protocol: string(s.Protocol())})
+			}
+		}
+		sort.Sort(&ci)
+		out.Peers = append(out.Peers, ci)
+	}
+
+	sort.Sort(&out)
+	return &out, nil
 }
